show repl locals
