+++ public/diagram/diagram.css
@@ -1,1 +1,8 @@
 /* place your styles here */
+.button {
+  border-radius: 15%;
+  background-color: white;
+  padding: 0.5em 0.2em;
+  font-family: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif;
+  font-weight: 600;
+}
