+++ tsconfig.build.json
@@ -1,4 +1,7 @@
 {
   "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "sourceMap": true
+  },
   "include": ["src/**/*.ts"]
 }
